# specs/001-ai-study-generator/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI-assisted Study Generator API
  version: 0.1.0
servers:
  - url: /api
paths:
  /auth/signup:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                displayName:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Created
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
  /courses:
    post:
      summary: Create a course
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                emoji:
                  type: string
                sourceText:
                  type: string
      responses:
        '201':
          description: Created
    get:
      summary: List courses for authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /courses/{courseId}/summaries:
    post:
      summary: Generate or regenerate a summary for a course
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                regenerate:
                  type: boolean
      responses:
        '201':
          description: Summary created
  /courses/{courseId}/quizzes:
    post:
      summary: Generate a quiz from a course or summary
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                count:
                  type: integer
                type:
                  type: string
      responses:
        '201':
          description: Quiz created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizResponse'
    get:
      summary: List quizzes for a course
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizListResponse'
  /quizzes/{quizId}:
    get:
      summary: Get quiz by id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: quizId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizResponse'
  /quizzes/{quizId}/attempts:
    post:
      summary: Submit quiz answers and get score
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: quizId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitQuizAttemptRequest'
      responses:
        '201':
          description: Attempt created with score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizAttemptResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    QuizQuestionMCQ:
      type: object
      required: [id, type, question, options, correctAnswer, explanation]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [MCQ]
        question:
          type: string
        options:
          type: array
          minItems: 4
          maxItems: 4
          items:
            type: string
        correctAnswer:
          type: integer
          minimum: 0
          maximum: 3
        explanation:
          type: string
    QuizQuestionOPEN:
      type: object
      required: [id, type, question, correctAnswer, explanation]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [OPEN]
        question:
          type: string
        correctAnswer:
          type: string
        explanation:
          type: string
    QuizQuestion:
      oneOf:
        - $ref: '#/components/schemas/QuizQuestionMCQ'
        - $ref: '#/components/schemas/QuizQuestionOPEN'
      discriminator:
        propertyName: type
        mapping:
          MCQ: '#/components/schemas/QuizQuestionMCQ'
          OPEN: '#/components/schemas/QuizQuestionOPEN'
    QuizParams:
      type: object
      required: [count, types]
      properties:
        count:
          type: integer
        types:
          type: array
          items:
            type: string
    QuizResponse:
      type: object
      required: [id, courseId, params, questions, createdAt]
      properties:
        id:
          type: string
        courseId:
          type: string
        params:
          $ref: '#/components/schemas/QuizParams'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuizQuestion'
        createdAt:
          type: string
          format: date-time
        lastAttemptSummary:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/LastAttemptSummary'
    QuizListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QuizResponse'
    LastAttemptSummary:
      type: object
      required: [attemptId, score, submittedAt]
      properties:
        attemptId:
          type: string
        score:
          type: integer
        submittedAt:
          type: string
          format: date-time
    SubmitQuizAttemptRequest:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: object
            required: [questionId, answer]
            properties:
              questionId:
                type: string
              answer:
                oneOf:
                  - type: integer
                    minimum: 0
                    maximum: 3
                  - type: string
    QuizAttemptResponse:
      type: object
      required: [attemptId, quizId, userId, score, submittedAt]
      properties:
        attemptId:
          type: string
        quizId:
          type: string
        userId:
          type: string
        score:
          type: integer
        submittedAt:
          type: string
          format: date-time